removed synchronous code
