+++ airbase/__init__.py
@@ -1,3 +1,2 @@
 from __future__ import absolute_import
 from .airtable import Airtable  # noqa: F401
-from .airtable_async import Airtable as AirtableAsync  # noqa: F401
