--- airbase/session/__init__.py
@@ -1,2 +1,0 @@
-from __future__ import absolute_import
-from .session import Session  # noqa: F401
