+++ airbase/tools.py
@@ -1,6 +1,8 @@
 from __future__ import absolute_import
 
-import copy
+import pandas as pd
+
+from typing import List
 
 from .utils import Logger
 
@@ -8,7 +10,133 @@ from .utils import Logger
 logger = Logger.start(__name__)
 
 
-def is_record(value):
+async def compare(
+    df_1: pd.DataFrame, df_2: pd.DataFrame, primary_keys: List[str]
+) -> pd.DataFrame:
+    """
+    Compare two pandas DataFrames and return a DataFrame with rows sorted by CRUD operation
+
+    Args:
+        df_1 (``pd.DataFrame``): DataFrame to compare (i.e. new payload)
+        df_2 (``pd.DataFrame``): DataFrame to compare against (i.e. existing data)
+        primary_keys(``list``): List of `str`` of the name of the primary keys to join the DataFrames
+
+    Returns:
+        pd.DataFrame
+    """  # noqa: E501
+
+    # list of headers
+    df_1_headers: List[str] = df_1.columns.values.tolist()
+    df_2_headers: List[str] = df_2.columns.values.tolist()
+
+    # combined and overlapping headers for dropping columns later
+    combined_headers: set = set(df_1_headers) | set(df_2_headers)
+    overlapping_headers: set = set(df_1_headers) & set(df_2_headers)
+    # list of overlapping_headers without primary keys
+    reduced_headers_a: List[str] = list(
+        overlapping_headers - set(primary_keys)
+    )
+    # list of combined_headers without primary keys
+    reduced_headers_b: List[str] = list(combined_headers - set(primary_keys))
+
+    # full outer join of both DataFrames on the primary_keys
+    # where the left is df_1 (with suffix '_x' applied to its column names)
+    #  and the right is df_2 (with suffix '_y' applied to its column names)
+    combined_df: pd.DataFrame = df_1.merge(
+        df_2, how="outer", on=primary_keys, indicator=True
+    )
+
+    # rows to be created are found by looking at unique rows in df_1 (left)
+    create_df: pd.DataFrame = (
+        combined_df.loc[lambda x: x["_merge"] == "left_only"]
+        .drop(
+            columns=[f"{other_header}_y" for other_header in reduced_headers_a]
+        )  # df_2 columns are dropped
+        .rename(
+            columns={
+                f"{other_header}_x": other_header
+                for other_header in reduced_headers_a
+            }
+        )  # df_1 columns are renamed back to original (without suffix)
+        .drop(columns=["_merge"])  # _merge indicator column is dropped
+    )
+
+    # rows to be deleted are found by looking at unique rows in df_2 (right)
+    delete_df: pd.DataFrame = (
+        combined_df.loc[lambda x: x["_merge"] == "right_only"]
+        .drop(
+            columns=[f"{other_header}_x" for other_header in reduced_headers_a]
+        )  # df_1 columns are dropped
+        .rename(
+            columns={
+                f"{other_header}_y": other_header
+                for other_header in reduced_headers_a
+            }
+        )  # df_1 columns are renamed back to original (without suffix)
+        .drop(columns=["_merge"])  # _merge indicator column is dropped
+    )
+
+    # rows to be updated are found by looking at rows in df_1 (left)
+    # that differ from df_2 (right) minus the rows to be created (create_df)
+
+    # full outer join of both DataFrames on no keys
+    # where the left is df_1 (with suffix '_x' applied to its column names)
+    #  and the right is df_2 (with suffix '_y' applied to its column names)
+    df_1_unique_rows: pd.DataFrame = (
+        df_1.merge(df_2, how="outer", indicator=True, sort=True)
+        .loc[
+            lambda x: x["_merge"] == "left_only"
+        ]  # only df_1 unique rows are kept
+        .drop(columns=["_merge"])  # _merge indicator column is dropped
+    )
+
+    # full outer join of both DataFrames on no primary keys
+    # where the left is df_1_unique_rows
+    # (with suffix '_x' applied to its column names)
+    # and the right is create_df
+    # (with suffix '_y' applied to its column names)
+    update_df: pd.DataFrame = (
+        df_1_unique_rows.merge(
+            create_df, indicator=True, how="outer", on=primary_keys
+        )
+        .loc[
+            lambda x: x["_merge"] != "both"
+        ]  # only df_1_unique_rows unique rows are kept
+        .drop(
+            columns=[f"{other_header}_y" for other_header in reduced_headers_b]
+        )  # create_df columns are dropped
+        .rename(
+            columns={
+                f"{other_header}_x": other_header
+                for other_header in reduced_headers_b
+            }
+        )  # df_1_unique_rows columns are renamed back to original
+        .drop(columns=["_merge"])  # _merge indicator column is dropped
+    )
+
+    # insert crud_type for each
+    create_df.insert(
+        loc=len(df_1_headers),
+        column="crud_type",
+        value=["create"] * create_df.shape[0],
+    )
+
+    update_df.insert(
+        loc=len(df_1_headers),
+        column="crud_type",
+        value=["update"] * update_df.shape[0],
+    )
+
+    delete_df.insert(
+        loc=len(df_2_headers),
+        column="crud_type",
+        value=["delete"] * delete_df.shape[0],
+    )
+
+    return create_df.append(update_df, sort=True).append(delete_df, sort=True)
+
+
+async def is_record(value):
     """
     Checks whether a value is a Record ID or a list of Record IDs
 
@@ -22,7 +150,7 @@ def is_record(value):
     return isinstance(value, str) and value[0:3] == "rec" and len(value) == 17
 
 
-def get_primary_keys_as_hashable(record, primary_keys):
+async def get_primary_keys_as_hashable(record, primary_keys):
     hashable_keys = []
     for key in primary_keys:
         val = record["fields"].get(key)
@@ -33,7 +161,7 @@ def get_primary_keys_as_hashable(record, primary_keys):
     return tuple(hashable_keys) if hashable_keys else None
 
 
-def graft_fields(record, fields, separator=",", sort=True):
+async def graft_fields(record, fields, separator=",", sort=True):
 
     for field in fields:
         value = record["fields"].get(field)
@@ -48,7 +176,7 @@ def graft_fields(record, fields, separator=",", sort=True):
     return record
 
 
-def link_tables(
+async def link_tables(
     table_a, table_b, fields_to_link_in_a, primary_key_b,
 ):
     """
@@ -69,9 +197,9 @@ def link_tables(
         if record_b["fields"].get(primary_key_b)
     }
 
-    new_table = []
+    # new_table = []
     for record_a in table_a:
-        new_record = copy.deepcopy(record_a)
+        # new_record = copy.deepcopy(record_a)
         for field_to_link in fields_to_link_in_a:
             field_to_link = field_to_link.strip()
             val = record_a["fields"][field_to_link]
@@ -79,16 +207,16 @@ def link_tables(
                 continue
 
             keys = (x.strip() for x in val.split(","))
-            new_record["fields"][field_to_link] = [
+            record_a["fields"][field_to_link] = [
                 table_b_by_primary_key.get(key)
                 for key in keys
                 if table_b_by_primary_key.get(key)
             ]
-        new_table.append(new_record)
-    return new_table
+        # new_table.append(new_record)
+    return table_a
 
 
-def combine_records(record_a, record_b, join_fields=None):
+async def combine_records(record_a, record_b, join_fields=None):
     """
     Combines unique information from two records into 1.
 
@@ -128,7 +256,7 @@ def combine_records(record_a, record_b, join_fields=None):
         return record_a
 
 
-def filter_record(record_a, record_b, filter_fields=None):
+async def filter_record(record_a, record_b, filter_fields=None):
     """
     Filters a record for unique information.
 
@@ -160,7 +288,7 @@ def filter_record(record_a, record_b, filter_fields=None):
     return record
 
 
-def override_record(record, existing_record, overrides):
+async def override_record(record, existing_record, overrides):
     """
     Removes fields from record if user has overriden them on airtable.
 
@@ -183,7 +311,7 @@ def override_record(record, existing_record, overrides):
     return record
 
 
-def compare_records(
+async def compare_records(
     record_a, record_b, method, overrides=None, filter_fields=None
 ):
     """
@@ -217,7 +345,7 @@ def compare_records(
         logger.warning("Invalid record format provided.")
 
 
-def replace_values(field, value):
+async def replace_values(field, value):
     # Simplify attachement objects
     if isinstance(value, list) and isinstance(value[0], dict):
         new_value = [{"url": obj["url"]} for obj in value if "url" in obj]
